moved ball to its own module
